Level, Obj Save/Load (1)
- Level과 Spring ,platform, Strawberry, Zipmover객체의 Save Load구현 (binary로)
- Load되는 거 확인은 했는데 프로그램 종료할 때 tRoom의 vector<Vec2>의 소멸자 호출하다가 예외 발생하는듯

+++ CTileMap.cpp
@@ -6,8 +6,8 @@
 #include "CTexture.h"
 
 CTileMap::CTileMap()
-	: m_UnitWidth(TILE_SCALE)
-	, m_UnitHeight(TILE_SCALE)
+	: m_UnitWidth((UINT)TILE_SCALE)
+	, m_UnitHeight((UINT)TILE_SCALE)
 	, m_RowCnt(0)
 	, m_ColCnt(0)
 {
@@ -51,7 +51,7 @@ void CTileMap::AddCollider()
 				continue;
 			}
 
-			Vec2 vTilePos = Vec2(Col * m_UnitWidth, Row * m_UnitHeight) + m_Offset;
+			Vec2 vTilePos = Vec2((float)Col * m_UnitWidth, (float)Row * m_UnitHeight) + m_Offset;
 
 			const tColInfo& Info = m_vecTile[Row * m_ColCnt + Col]->GetColInfo();
